wip: update t16 branch (#199)
* improve extendable openapi definition #168
* cleanup
* dapa: correct link to variables
* dapa: fix grid resampling
* dapa: support csv and tiff
* dapa: add html for the dapa page
* dapa: support for coordRef and other improvements

+++ ogcapi-draft/ogcapi-observation-processing/src/main/java/de/ii/ldproxy/ogcapi/observation_processing/data/GeometryPoint.java
@@ -5,31 +5,31 @@ import com.google.common.collect.ImmutableList;
 import java.util.List;
 
 public class GeometryPoint implements Geometry {
-    final List<Float> coord;
+    final List<Double> coord;
 
-    public GeometryPoint(List<Float> coord) {
+    public GeometryPoint(List<Double> coord) {
         this.coord = coord;
     }
 
-    public GeometryPoint(float lon, float lat) {
+    public GeometryPoint(double lon, double lat) {
         this.coord = ImmutableList.of(lon, lat);
     }
 
-    public GeometryPoint(float lon, float lat, float alt) {
+    public GeometryPoint(double lon, double lat, double alt) {
         this.coord = ImmutableList.of(lon, lat, alt);
     }
 
-    public float getLon() {
-        return coord.size()>=1 ? coord.get(0) : Float.NaN;
+    public Double getLon() {
+        return coord.size()>=1 ? coord.get(0) : Double.NaN;
     }
 
-    public float getLat() {
-        return coord.size()>=2 ? coord.get(1) : Float.NaN;
+    public Double getLat() {
+        return coord.size()>=2 ? coord.get(1) : Double.NaN;
     }
 
-    public float getAlt() {
-        return coord.size()>=3 ? coord.get(2) : Float.NaN;
+    public Double getAlt() {
+        return coord.size()>=3 ? coord.get(2) : Double.NaN;
     }
 
-    public List<Float> asList() { return coord; }
+    public List<Double> asList() { return coord; }
 }
